refactor: move SingularityReader to own file
